Fixes: drivers/net/can/spi/mcp25xxfd: correct wordwise readout of SPI data from SRAM
Inserted a module parameter to read SRAM data wordwise as specified in the mcp2517fd
datasheet. This parameter is set default to true and can be used to change back to
the original handling if allowed by later chip generations.
Adapted the RX prefetch and receive handling in mcp25xxfd_can_rx_read_frame() to read
wordwise instead of bytewise if configured in the module parameter. This fixes the
problem of junk data read on frame reception after a prefetch is done that's not
wordwise.