drivers/net/can/spi/mcp25xxfd: clear TX FIFOs on interface shutdown to avoid send errors/exceptions
Added a reset on interface stop in mcp25xxfd_can_stop() and cleared TEF and submit
queue in order to stop all pending send operations. Without the fix, we sometimes
observed exceptions because no TX FIFO was left to send on as it was occupied and
never freed, as well as a bulk send of data after reactivating data from previously
configured TEFs.