drivers/net/can/spi/mcp25xxfd: workaround for not supported IMVIF situation on empty bdiag register
In some situations with bad timing, the IVMIF interrupt fires without any of the
checked BDIAG1 bits set. This previously lead to a negative response code and in turn
the deactivation of all interrupts for the driver through the kernel. Unfortunately,
we were not able to find the root cause of the error, but it seems to happen if no
CAN connection is present at and interface and a specific timing behavior takes place.

Because we don't want the system to stop working in such conditions, in
mcp25xxfd_can_int_handle_ivmif() it's checked if the condition occured, that means if
all the checked flags are not set while the IVMIF interrupt occured. Instead of
returning an error code, in such cases a warning is issued into the kernel message
log. If any of the not checked flags, that are currently reserved in the datasheet,
may be set, the old behavior is kept so if later a switch to a new generation of the
chip is used and a new flag is used.
Also, the missing evaluation of the DLCMM flag in BDIAG1 has been added to
mcp25xxfd_can_int_handle_ivmif_rx().